Deprecate legacy bookmaker utilities
